Minimise dependencies
The WASM demo was built using Webpack, which has about 450
subdependencies and does not even work with newer versions of Node.js.
This commit simplifies things by removing Webpack, which turned out to
be unnecessary anyway.

As part of the simplification, I had to remove the Conway's Game of Life
stuff, which, until I realised it just was the remnant of a tutorial,
made the demo quite confusing.