one-time host initialization controlled by `prepared`
The initial host memory image contains a single event
and an empty continuation queue. The event creates a new
continuation, starting execution of the bootstrap actor.
Unfortunately, the GUI is unable to display the state of
the VM when there is no "current continuation", therefore
we have to "prepare" the VM before was can view the state.