Implement `actor_revert` and `event_inject`
The _revert_ functionality takes the current continuation
and returns the configuration to the state before the
associated event was dispatched. The actor state is reset
and the event is injected at the front of the event queue
for immediate dispatch. The _inject_ functionality will
also be used by interrupts, to improve latency.