Stricter assembler and loader
This commit adds type checking for instruction streams and dictionaries,
ensuring that an instruction does not specify a data value as its continuation.
It also adds checks for cyclic refs, lists and dictionaries.

It fixes a bug in the parser whereby the input object might have been corrupted
by backtracking.

It is now possible to omit the definitions and exports in a module, meaning an
empty file will assemble to an empty CRLF object.