Standardize 'beh' export
It is desirable to refer to the main behavior locally like

    push cell_beh   ; e.g. in cell.asm

but refer to the behavior externally like

    push cell.beh   ; e.g. in serial.asm

rather than

    push cell.cell_beh

which is repetitive and, worse, implies that the cell.asm module exports
behaviors besides 'cell_beh'.

I propose that we alias the main behavior in a module as 'beh' and export just
the alias.