improve efficiency and security of WebRTC signalling
Previously, signalling was symmetric. Connecting and listening parties both
used the signalling server in the same way - by subscribing to signals destined
for a name. The signalling server blindly trusted subscribing parties to give
their real name, making it trivial to eavesdrop on signals intended for others.
While this was not a security risk per se, it did make a DoS attack trivial to
implement. (This approach was also very inefficient for listening parties,
because often a bunch of identical connections would be made to the signalling
server and they would all receive the same signals.)

While there is no easy way for the signalling server to verify a party's name
(because a name is an RTCCertificate fingerprint), it is possible to demand
credentials from parties attempting to listen while providing an unhindered
endpoint for parties attempting to connect. (The credentials are a static
password, "uFork", until we come up with something better.)

This commit introduces asymmetric signalling: there are now different endpoints
for connecting and listening. I also managed to save a round trip by encoding
the SDP offer in a URL, which is nice.