remove UNIT from machines
This commit removes the UNIT value from the VM and uCode implementations of
uFork.

The area of reserved ROM previously occupied by UNIT is now marked as unused.

The VM's garbage collector previously used UNIT to represent the "black" color
and UNDEF as the "white" color. These have been remapped with TRUE as "black"
(reachable) and FALSE as "white" (unscanned).