avoid 'my beh' where possible
The 'my beh' instruction is very powerful. It lets an actor update its state
without knowing its behavior (code). In most cases, however, this is not
necessary as the behavior can be referenced by name. Doing so makes the code
more statically analyzable, preventing a maintenance hazard that arises when
attempting to change the state signature of a behavior.