//debug -> import.meta.main
Previously, our JavaScript module demos were disabled in production by
prefixing each line with //debug, and enabled in Replete by preprocessing
evaluated source code with Ecomcon.

The benefit of the Ecomcon approach was that stripping the demo from the
source, for distribution, would have been trivially easy (had we bothered to
strip the demos, which we didn't). The drawbacks were that:

- we needed an editor plugin to easily add or remove the //debug prefix
- the demos could only be run using Replete (yet another editor plugin)
- syntax highlighting was unavailable whilst the //debug prefix was present
- it looked ugly and strange

This commit replaces the Ecomcon approach with a runtime check of
'import.meta.main': true when the module is being run as the entry point (or
evaluated in Replete) and falsy otherwise. The benefit is that no editor
support is required, and it looks much less strange. The drawback is that a
more sophisticated build tool will be required to strip the modules of their
demos, as might be desirable for distribution or bundling.