remove 'my beh' instruction
The 'my beh' instruction allowed an actor to obtain a reference to its own
code. To date, we have not found a good use case for this: it is better to
statically reference behaviors, by name, than to reference them dynamically via
'my beh'. We have never encounted a case where the current behavior was not
known ahead of time.