() -> #nil
This commit removes () from our stack diagrams and documentation, replacing it
with #nil. () remains a valid representation of NIL in Scheme and Humus
programs, but nowhere else.

This is the first step in the migration from Scheme-style to Humus-style list
depictions, for example (a b c) -> a,b,c,#nil.