actor graph view for debugger
This commit adds a view to the debugger that renders a core's actors in
a force-directed graph, animating changes as they happen. The graph is
drawn on a <canvas> and is currently very basic, showing only actors and
the references between them.

Some keyboard shortcuts were added to the debugger for changing views:
alt+up and alt+down. The playback interval slider was inverted to become
a speed slider, and turned from linear to logarithmic.