move ROM from playground to debugger
The debugger needed a ROM viewer, and I thought it fit better in the
debugger than the playground because ROM is only populated once the core
is running. It is still accessible from the playground via the debugger.
Eventually, the debugger and playground will become tightly integrated
and it will be easier to get to the ROM viewer from the playground.

The "disassembly" feature of the ROM viewer was moved out of the ROM
viewer and into the playground, because it has nothing to do with ROM
and should update in realtime as the text changes.