make handle_event return an event pointer
Previously, it was not possible associate a device with its synchronous
effects. This commit changes the signature of 'handle_event' such that
it may inform the core what event, if any, it generated in response to a
message. Currently there is a limit of 1 generated event, but that could
be changed if necessary.

The on_trace function has been renamed on_txn, and is no longer called
on 'end stop' or 'end abort' instructions.