reduce default sponsor limits
Previously, we set the root sponsor limits to the maximum positive
fixnum for a 32-bit architecture. This allowed the core to run for
several minutes without refilling, which resulting in the auto
refilling feature rarely being exercised.

Now we set it to the maximum positive fixnum for a 16-bit architecture,
several magnitudes smaller.